+++ opensrp-chw-core/src/main/java/org/smartregister/chw/core/contract/AncMemberProfileContract.java
@@ -1,7 +1,8 @@
-package org.smartregister.chw.hf.contract;
+package org.smartregister.chw.core.contract;
 
 import org.smartregister.chw.anc.contract.BaseAncMemberProfileContract;
 import org.smartregister.domain.Task;
+import org.smartregister.repository.AllSharedPreferences;
 
 import java.util.Set;
 
@@ -15,9 +16,13 @@ public interface AncMemberProfileContract {
         void fetchTasks();
 
         void setEntityId(String entityId);
+
+        void createSickChildEvent(AllSharedPreferences allSharedPreferences, String jsonString) throws Exception;
     }
 
     interface Interactor extends BaseAncMemberProfileContract.Interactor {
+        void createSickChildEvent(AllSharedPreferences allSharedPreferences, String jsonString, String entityID) throws Exception;
+
         void getClientTasks(String planId, String baseEntityId, AncMemberProfileContract.InteractorCallBack callback);
     }
