+++ opensrp-chw-core/src/main/java/org/smartregister/chw/core/interactor/CoreAncMemberProfileInteractor.java
@@ -3,6 +3,8 @@ package org.smartregister.chw.core.interactor;
 import android.content.Context;
 
 import org.ei.drishti.dto.AlertStatus;
+import org.jetbrains.annotations.NotNull;
+import org.joda.time.DateTime;
 import org.smartregister.chw.anc.AncLibrary;
 import org.smartregister.chw.anc.contract.BaseAncMemberProfileContract;
 import org.smartregister.chw.anc.domain.MemberObject;
@@ -10,10 +12,18 @@ import org.smartregister.chw.anc.domain.Visit;
 import org.smartregister.chw.anc.interactor.BaseAncMemberProfileInteractor;
 import org.smartregister.chw.anc.util.Constants;
 import org.smartregister.chw.core.R;
+import org.smartregister.chw.core.application.CoreChwApplication;
+import org.smartregister.chw.core.contract.AncMemberProfileContract;
+import org.smartregister.chw.core.utils.CoreConstants;
+import org.smartregister.chw.core.utils.CoreReferralUtils;
+import org.smartregister.domain.Task;
+import org.smartregister.repository.AllSharedPreferences;
+import org.smartregister.repository.BaseRepository;
 
 import java.util.Date;
+import java.util.Set;
 
-public class CoreAncMemberProfileInteractor extends BaseAncMemberProfileInteractor {
+public class CoreAncMemberProfileInteractor extends BaseAncMemberProfileInteractor implements AncMemberProfileContract.Interactor {
     private Context context;
 
     public CoreAncMemberProfileInteractor(Context context) {
@@ -54,4 +64,33 @@ public class CoreAncMemberProfileInteractor extends BaseAncMemberProfileInteract
         return lastVisitDate;
     }
 
+    @Override
+    public void createSickChildEvent(AllSharedPreferences allSharedPreferences, String jsonString, String entityId) throws Exception {
+        CoreReferralUtils.createReferralEvent(allSharedPreferences, jsonString, CoreConstants.TABLE_NAME.ANC_REFERRAL, entityId);
+    }
+
+    @Override
+    public void getClientTasks(String planId, String baseEntityId, @NotNull AncMemberProfileContract.InteractorCallBack callback) {
+        Set<Task> taskList = CoreChwApplication.getInstance().getTaskRepository().getTasksByEntityAndStatus(planId, baseEntityId, Task.TaskStatus.READY);
+        Task task = new Task();
+        task.setIdentifier("iudsfsigdfdsyud");
+        task.setFocus("ANC Referral");
+        task.setGroupIdentifier("iudsfsigdfdsyud");
+        task.setStatus(Task.TaskStatus.READY);
+        task.setBusinessStatus(CoreConstants.BUSINESS_STATUS.REFERRED);
+        task.setPriority(3);
+        task.setCode("Referral");
+        task.setDescription("Review and perform the referral for the client"); //set to string
+        task.setForEntity(baseEntityId);
+        DateTime now = new DateTime();
+        task.setExecutionStartDate(now);
+        task.setAuthoredOn(now);
+        task.setLastModified(now);
+        task.setOwner("iudsfsigdfdsyud");
+        task.setSyncStatus(BaseRepository.TYPE_Created);
+        task.setRequester("iudsfsigdfdsyud");
+        task.setLocation("iudsfsigdfdsyud");
+        taskList.add(task);
+        callback.setClientTasks(taskList);
+    }
 }
