+++ opensrp-chw-core/src/main/java/org/smartregister/chw/core/utils/CoreReferralUtils.java
@@ -1,9 +1,22 @@
 package org.smartregister.chw.core.utils;
 
+import org.joda.time.DateTime;
+import org.json.JSONException;
+import org.json.JSONObject;
+import org.smartregister.chw.anc.util.NCUtils;
+import org.smartregister.chw.core.application.CoreChwApplication;
+import org.smartregister.clientandeventmodel.Event;
 import org.smartregister.cursoradapter.SmartRegisterQueryBuilder;
+import org.smartregister.domain.Task;
 import org.smartregister.family.util.DBConstants;
+import org.smartregister.location.helper.LocationHelper;
+import org.smartregister.repository.AllSharedPreferences;
+import org.smartregister.repository.BaseRepository;
 
 import java.util.ArrayList;
+import java.util.UUID;
+
+import timber.log.Timber;
 
 public class CoreReferralUtils {
     public static String mainSelect(String tableName, String familyTableName, String familyMemberTableName, String mainCondition) {
@@ -38,5 +51,60 @@ public class CoreReferralUtils {
         columnList.add(tableName + "." + DBConstants.KEY.DOB);
         columnList.add(tableName + "." + org.smartregister.family.util.Constants.JSON_FORM_KEY.DOB_UNKNOWN);
         return columnList.toArray(new String[columnList.size()]);
+    }
+
+    public static void createReferralEvent(AllSharedPreferences allSharedPreferences, String jsonString, String referralTable, String entityId) throws Exception {
+        final Event baseEvent = org.smartregister.chw.anc.util.JsonFormUtils.processJsonForm(allSharedPreferences, setEntityId(jsonString, entityId), referralTable);
+        NCUtils.processEvent(baseEvent.getBaseEntityId(), new JSONObject(org.smartregister.chw.anc.util.JsonFormUtils.gson.toJson(baseEvent)));
+        createReferralTask(baseEvent.getBaseEntityId(), allSharedPreferences);
+    }
+
+    private static String setEntityId(String jsonString, String entityId) {
+        String referralForm = "";
+        try {
+            JSONObject jsonObject = new JSONObject(jsonString);
+            jsonObject.put(CoreConstants.ENTITY_ID, entityId);
+
+            referralForm = jsonObject.toString();
+        } catch (JSONException e) {
+            Timber.e(e, "CoreChildProfileInteractor --> setEntityId");
+        }
+
+        return referralForm;
+    }
+
+    private static void createReferralTask(String baseEntityId, AllSharedPreferences allSharedPreferences) {
+        Task task = new Task();
+        task.setIdentifier(UUID.randomUUID().toString());
+        //TODO Implement plans
+      /*  Iterator<String> iterator = ChwApplication.getInstance().getPlanDefinitionRepository()
+                .findAllPlanDefinitionIds().iterator();
+        if (iterator.hasNext()) {
+            task.setPlanIdentifier(iterator.next());
+        } else {
+
+            Timber.e("No plans exist in the server");
+        }*/
+        task.setPlanIdentifier("5270285b-5a3b-4647-b772-c0b3c52e2b71");
+        LocationHelper locationHelper = LocationHelper.getInstance();
+        ArrayList<String> allowedLevels = new ArrayList<>();
+        allowedLevels.add(CoreConstants.CONFIGURATION.HEALTH_FACILITY_TAG);
+        task.setGroupIdentifier(locationHelper.getOpenMrsLocationId(locationHelper.generateDefaultLocationHierarchy(allowedLevels).get(0)));
+        task.setStatus(Task.TaskStatus.READY);
+        task.setBusinessStatus(CoreConstants.BUSINESS_STATUS.REFERRED);
+        task.setPriority(3);
+        task.setCode("Referral");
+        task.setDescription("Review and perform the referral for the client"); //set to string
+        task.setFocus(CoreConstants.TASKS_FOCUS.SICK_CHILD);//the same here
+        task.setForEntity(baseEntityId);
+        DateTime now = new DateTime();
+        task.setExecutionStartDate(now);
+        task.setAuthoredOn(now);
+        task.setLastModified(now);
+        task.setOwner(allSharedPreferences.fetchRegisteredANM());
+        task.setSyncStatus(BaseRepository.TYPE_Created);
+        task.setRequester(allSharedPreferences.getANMPreferredName(allSharedPreferences.fetchRegisteredANM()));
+        task.setLocation(allSharedPreferences.fetchUserLocalityId(allSharedPreferences.fetchRegisteredANM()));
+        CoreChwApplication.getInstance().getTaskRepository().addOrUpdate(task);
     }
 }
