+++ opensrp-chw-hf/src/main/assets/ec_client_classification.json
@@ -146,6 +146,13 @@
           },
           {
             "field": "eventType",
+            "field_value": "ANC Referral",
+            "creates_case": [
+              "ec_anc_referral"
+            ]
+          },
+          {
+            "field": "eventType",
             "field_value": "Close Referral",
             "creates_case": [
               "ec_close_referral"
