+++ opensrp-chw-hf/src/main/assets/ec_client_fields.json
@@ -1000,7 +1000,7 @@
           "type": "Event",
           "json_mapping": {
             "field": "obs.fieldCode",
-            "concept": ""
+            "formSubmissionField": "referral_task_previous_status"
           }
         }
       ]
