+++ opensrp-chw-hf/src/main/java/org/smartregister/chw/hf/utils/HfReferralUtils.java
@@ -1,0 +1,6 @@
+package org.smartregister.chw.hf.utils;
+
+import org.smartregister.chw.core.utils.CoreReferralUtils;
+
+public class HfReferralUtils extends CoreReferralUtils {
+}
