+++ opensrp-chw/src/ba/assets/ec_client_classification.json
@@ -143,6 +143,13 @@
             "creates_case": [
               "ec_child_referral"
             ]
+          },
+          {
+            "field": "eventType",
+            "field_value": "ANC Referral",
+            "creates_case": [
+              "ec_anc_referral"
+            ]
           }
         ]
       }
